chore(parser): move getOperationResponseCode to common parser
